requirements.txt
playwright==1.44.0
requests==2.32.3
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-api-python-client==2.131.0
